+++ packages/diat/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.2.0
 feat: support "attachConsole" in diat-node-inspect
 feat: support "setLogpoint" in diat-node-inspect
+feat: support "getScripts" in diat-node-inspect
+feat: support "scriptSource" in diat-node-inspect
 
 ## 1.1.6